feat: finish F2 decision claim coverage CLI
