+++ packages/cli/src/cli/command-input-descriptors.ts
@@ -65,6 +65,7 @@ const explicitInputDescriptors = {
       rejected: { type: ["string", "array"], description: "Rejected option text, or an array whose first item has text and whyNot.", items: { type: "object", properties: { text: { type: "string" }, whyNot: { type: "string" } } } },
       whyNot: { type: "string", description: "Rationale for rejecting the rejected option." },
       claim: { type: "string", description: "Optional supporting claim." },
+      claims: { type: "array", description: "Supporting claims born with the decision.", items: { type: "object", properties: { id: { type: "string" }, text: { type: "string" }, load_bearing: { type: "boolean" } } } },
       riskTier: { type: "string", description: "Decision risk tier: low, medium, or high." },
       urgency: { type: "string", description: "Decision urgency: low, medium, or high." },
       proposedBy: { type: "string", description: "Actor ref such as agent:codex." },
@@ -83,6 +84,7 @@ const explicitInputDescriptors = {
       shortcut("--rejected", "$.rejected", "set"),
       shortcut("--why-not", "$.whyNot", "set"),
       shortcut("--claim", "$.claim", "set"),
+      shortcut("--claim", "$.claims", "append"),
       shortcut("--risk-tier", "$.riskTier", "set"),
       shortcut("--urgency", "$.urgency", "set"),
       shortcut("--proposed-by", "$.proposedBy", "set"),
