+++ packages/cli/test/conflict-preflight-cli.test.ts
@@ -17,6 +17,7 @@ const expectedConflictPreflightKinds = [
   "decision-amend",
   "decision-defer",
   "decision-propose",
+  "decision-reckon",
   "decision-reject",
   "decision-relate",
   "decision-relation-replace",
@@ -72,6 +73,7 @@ const expectedAutoRuntimeEventKinds = [
   "decision-amend",
   "decision-defer",
   "decision-propose",
+  "decision-reckon",
   "decision-reject",
   "decision-relate",
   "decision-relation-replace",
