+++ packages/kernel/src/entity/field-contracts.ts
@@ -9,7 +9,7 @@ export type EntityFieldReadSurface =
   | { readonly kind: "projection"; readonly path: string; readonly queryable: boolean }
   | { readonly kind: "show"; readonly path: string };
 export type EntityFieldWriteSurface =
-  | { readonly kind: "amend"; readonly operation: "replace" | "append" }
+  | { readonly kind: "amend"; readonly operation: "replace" | "append" | "metadata" }
   | { readonly kind: "lifecycle"; readonly operation: string };
 
 export interface EntityFieldContract {
@@ -43,7 +43,7 @@ export const decisionFieldContracts = {
   question: immutable("changing the core question changes the decision identity; use supersede", projection("question", true), show("decision.question")),
   chosen: amendable([amendWrite("append")], projection("chosen", false), show("decision.chosen")),
   rejected: amendable([amendWrite("append")], projection("rejected", false), show("decision.rejected")),
-  claims: amendable([amendWrite("append"), amendWrite("replace")], show("decision.claims")),
+  claims: amendable([amendWrite("append"), amendWrite("metadata")], show("decision.claims")),
   relations: immutable("relation changes require relation/evidence-specific write surfaces", show("decision.relations"))
 } satisfies Record<DecisionFieldKey, EntityFieldContract>;
