+++ tools/test-tier-manifest.mjs
@@ -85,6 +85,7 @@ export const testTierManifest = {
     "packages/cli/test/check-governance-cli.test.ts",
     "packages/cli/test/conflict-preflight-cli.test.ts",
     "packages/cli/test/decision-cli.test.ts",
+    "packages/cli/test/decision-coverage-cli.test.ts",
     "packages/cli/test/docmap-cli.test.ts",
     "packages/cli/test/distill-cli.test.ts",
     "packages/cli/test/doctor-cli.test.ts",
@@ -98,6 +99,7 @@ export const testTierManifest = {
     "packages/cli/test/p16-command-parity-cli.test.ts",
     "packages/cli/test/post-merge-check-cli.test.ts",
     "packages/cli/test/projection-freshness-cli.test.ts",
+    "packages/cli/test/preset-milestone-closeout-cli.test.ts",
     "packages/cli/test/preset-module-cli.test.ts",
     "packages/cli/test/preset-script-cli.test.ts",
     "packages/cli/test/preset-script-dossier-cli.test.ts",
